+++ packages/esm-patient-appointments-app/translations/km.json
@@ -25,14 +25,14 @@
   "date": "កាលបរិច្ឆេទ",
   "dateTime": "កាលបរិច្ឆេទ និងពេលវេលាណាត់ជួប",
   "day": "ថ្ងៃ",
-  "daysOfWeek": "ថ្ងៃ\u200bនៃ\u200bស\u200bប្តា\u200bហ៍",
+  "daysOfWeek": "ថ្ងៃ​នៃ​ស​ប្តា​ហ៍",
   "discard": "បោះបង់",
   "durationInMinutes": "រយៈពេល (នាទី)",
   "edit": "កែសម្រួល",
   "editAppointment": "កែសម្រួលការណាត់ជួប",
   "endDate": "កាលបរិច្ឆេទបញ្ចប់",
   "invalidNumber": "លេខមិនត្រឹមត្រូវទេ",
-  "isRecurringAppointment": "តើ\u200bនេះ\u200bជា\u200bការ\u200bណាត់\u200bជួប\u200bច្រើនដងមែនទេ?",
+  "isRecurringAppointment": "តើ​នេះ​ជា​ការ​ណាត់​ជួប​ច្រើនដងមែនទេ?",
   "loading": "កំពុងផ្ទុក",
   "location": "ទីតាំង",
   "no": "ទេ",
@@ -53,7 +53,7 @@
   "selectLocation": "ជ្រើសរើសទីតាំង",
   "selectService": "ជ្រើសរើសសេវាកម្ម",
   "service": "សេវាកម្ម",
-  "startDate": "ថ្ងៃ\u200bចាប់ផ្តើម",
+  "startDate": "ថ្ងៃ​ចាប់ផ្តើម",
   "status": "ស្ថានភាព",
   "time": "ពេលវេលា",
   "today": "ថ្ងៃនេះ",
