+++ packages/esm-patient-banner-app/translations/km.json
@@ -2,8 +2,8 @@
   "actions": "ជ្រើសរើសមុខងារ",
   "activeVisit": "ការមកពិនិត្យជំងឺសកម្ម",
   "address": "អាស័យដ្ឋាន",
-  "address1": "អាសយដ្ឋាន\u200bជួរទី១",
-  "address2": "អ\u200bស\u200bយ\u200bដ្ឋាន\u200bជួរ\u200bទី 2",
+  "address1": "អាសយដ្ឋាន​ជួរទី១",
+  "address2": "អ​ស​យ​ដ្ឋាន​ជួរ​ទី 2",
   "city": "ទីក្រុង",
   "cityVillage": "ទីក្រុង",
   "contactDetails": "ព័ត៌មានលម្អិតទំនាក់ទំនង",
@@ -17,7 +17,7 @@
   "male": "ប្រុស",
   "other": "ផ្សេងៗ",
   "patientLists": "បញ្ជីអ្នកជំងឺ",
-  "postalCode": "លេខ\u200bកូដ\u200bតំបន់",
+  "postalCode": "លេខ​កូដ​តំបន់",
   "relationships": "ទំនាក់ទំនង",
   "seeMoreLists_other": "មើលបញ្ជី {{count}} ទៀត",
   "showDetails": "បង្ហាញលម្អិត",
