+++ packages/esm-patient-chart-app/translations/km.json
@@ -10,7 +10,7 @@
   "appointmentUpdate": "ធ្វើបច្ចុប្បន្នភាពការណាត់ជួបនាពេលខាងមុខបានជោគជ័យ",
   "Attachments dashboard": "ផ្ទាំងគ្រប់គ្រងឯកសារភ្ជាប់",
   "cancel": "បោះបង់",
-  "cancelActiveVisitConfirmation": "តើ​អ្នក​ពិតជា​ចង់​បោះបង់ការពិនិត្យជំងឺ​សកម្ម​នេះទេ?",
+  "cancelActiveVisitConfirmation": "តើ\u200bអ្នក\u200bពិតជា\u200bចង់\u200bបោះបង់ការពិនិត្យជំងឺ\u200bសកម្ម\u200bនេះទេ?",
   "cancellingVisit": "បោះបង់ការពិនិត្យជំងឺ",
   "cancelVisit": "បោះបង់ការជួបពិនិត្យជំងឺ",
   "cancelVisitExplainerMessage": "បោះបង់ការពិនិត្យជំងឺនេះនឹងលុបការមកពិនិត្យជំងឺដែលពាក់ព័ន្ធ។",
@@ -61,7 +61,7 @@
   "errorDeletingVisit": "មានបញ្ហាបច្ចេកទេសក្នុងការលុបការពិនិត្យជំងឺ",
   "errorEndingVisit": "មានបញ្ហាបច្ចេកទេសបញ្ចប់ការពិនិត្យជំងឺ",
   "errorOccuredDeletingVisit": "មានបញ្ហាបច្ចេកទេសនៅពេលលុបការពិនិត្យជំងឺ",
-  "errorUpdatingVisitDetails": "មានបញ្ហាបច្ចេកទេស​ក្នុង​ការ​ធ្វើ​បច្ចុប្បន្នភាព​ព័ត៌មាន​លម្អិត​នៃ​ការ​ពិនិត្យជំងឺ",
+  "errorUpdatingVisitDetails": "មានបញ្ហាបច្ចេកទេស\u200bក្នុង\u200bការ\u200bធ្វើ\u200bបច្ចុប្បន្នភាព\u200bព័ត៌មាន\u200bលម្អិត\u200bនៃ\u200bការ\u200bពិនិត្យជំងឺ",
   "errorWhenRestoringVisit": "មានបញ្ហាបច្ចេកទេសនៅពេលទាញមកវិញនូវ{{visit}}",
   "failedDeleting": "មិនអាចលុបបានទេ",
   "failedToLoadCurrentVisit": "ការទាញទិន្នន័យពិនិត្យជំងឺបច្ចុប្បន្នបរាជ័យ",
@@ -173,7 +173,7 @@
   "visitEndSuccessfully": "បានបញ្ចប់ការមកពិនិត្យជំងឺបច្ចុប្បន្នដោយជោគជ័យ",
   "visitLocation": "ទីតាំងការមកពិនិត្យជំងឺ",
   "visitNotRestored": "ការពិនិត្យជំងឺមិនអាចទាញមកវិញបានទេ",
-  "visitRestored": "ការពិនិត្យជំងឺ​បាន​ទាញមកវិញបាន",
+  "visitRestored": "ការពិនិត្យជំងឺ\u200bបាន\u200bទាញមកវិញបាន",
   "visitRestoredSuccessfully": "{{visit}} បានទាញមកវិញដោយជោគជ័យ",
   "visits": "មកពិនិត្យជំងឺ",
   "Visits": "មកពិនិត្យជំងឺ",
