+++ packages/esm-patient-common-lib/src/orders/index.ts
@@ -1,3 +1,4 @@
 export * from './useOrderBasket';
 export * from './postOrders';
+export * from './useOrders';
 export * from './types';
