+++ packages/esm-patient-orders-app/translations/he.json
@@ -1,16 +1,15 @@
 {
-  "actions": "Actions",
   "add": "Add",
   "addResults": "Add Results",
   "cancelOrder": "Cancel Order",
   "dateOfOrder": "Date of Order",
   "female": "Female",
   "male": "Male",
-  "medications": "תרופות",
   "onDate": "on",
   "order": "Order",
   "orderedBy": "Ordered By",
   "orderId": "Order ID",
+  "orders": "Orders",
   "Orders": "Orders",
   "orderType": "Order Type",
   "other": "Other",
